+++ package.json
@@ -10,12 +10,16 @@
   "scripts": {
     "build": "turbo build",
     "ci": "npm run format && npm run build && npm run lint && npm run typecheck && npm run test:coverage",
+    "db:migrate": "tsx ./packages/openworkflow/postgres/scripts/db-migrate.ts",
+    "db:reset": "tsx ./packages/openworkflow/postgres/scripts/db-reset.ts",
     "format": "prettier --check . --ignore-path .gitignore --ignore-path .prettierignore",
     "format:fix": "prettier --write . --ignore-path .gitignore --ignore-path .prettierignore",
     "knip": "knip",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "ow": "tsx ./packages/cli/cli.ts",
+    "pghero": "sh ./packages/openworkflow/postgres/scripts/pghero.sh",
+    "squawk": "tsx ./packages/openworkflow/postgres/scripts/squawk.ts",
     "test": "vitest run",
     "test:coverage": "vitest run --coverage",
     "test:coverage:browse": "open ./coverage/index.html",
@@ -24,7 +28,6 @@
     "worker": "npx @openworkflow/cli worker start"
   },
   "dependencies": {
-    "@openworkflow/backend-postgres": "*",
     "openworkflow": "*"
   },
   "devDependencies": {
