--- packages/backend-postgres/.squawk.toml
@@ -1,4 +1,0 @@
-excluded_rules = [
-    "prefer-bigint-over-smallint",
-    "require-timeout-settings",
-]
