+++ packages/backend-postgres/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @openworkflow/backend-postgres
 
+## Unreleased
+
+- This package is now a thin compatibility shim. Prefer importing
+  `BackendPostgres` from `openworkflow/postgres`.
+
 ## 0.5.1
 
 - Change listWorkflowRuns to DESC ordering
