+++ packages/docs/docs.json
@@ -55,7 +55,13 @@
       },
       {
         "group": "Resources",
-        "pages": ["docs/roadmap", "docs/llms", "npm", "github"]
+        "pages": [
+          "docs/migration-v6",
+          "docs/roadmap",
+          "docs/llms",
+          "npm",
+          "github"
+        ]
       }
     ]
   },
