+++ packages/openworkflow/client.test.ts
@@ -1,6 +1,6 @@
-import { BackendPostgres } from "../backend-postgres/backend.js";
-import { DEFAULT_POSTGRES_URL } from "../backend-postgres/postgres.js";
 import { OpenWorkflow } from "./client.js";
+import { BackendPostgres } from "./postgres.js";
+import { DEFAULT_POSTGRES_URL } from "./postgres/postgres.js";
 import { defineWorkflowSpec } from "./workflow.js";
 import { type as arkType } from "arktype";
 import { randomUUID } from "node:crypto";
